Int to long!
Success (I should have thought to test this WAY earlier...

+++ BT Mark 3/app/src/main/java/com/example/biblio_tech_mark_3/Book.java
@@ -15,7 +15,7 @@ public class Book { //title, author, List genres, List Subjects, int ISBN, longS
     private List<String> genres; //More than one?
     @SerializedName("subjects")
     private List<String> subjects; //~6+2=8 long    entries__subjects__004, entries__subjects__005, entries__subjects__006, entries__subjects__001, entries__subjects__002,	entries__subjects__003 [subject_places], [subject_people];
-    private int ISBN; //The book's ISBN13 (not edition key, unique to book=edition). If ISBN10 then add prefix
+    private long ISBN; //The book's ISBN13 (not edition key, unique to book=edition). If ISBN10 then add prefix
     @SerializedName("description")
     private String notes; // also notes entries__first_sentence__value, entries__description, entries__description__value; //Have each value concatenate, not List?
 //    private boolean owned;
@@ -62,10 +62,10 @@ public class Book { //title, author, List genres, List Subjects, int ISBN, longS
     }
     public void setNotes(String notes) { this.notes = notes; }
 
-    public int getISBN() { return ISBN; }
-    public void setISBN(int ISBN) { this.ISBN = ISBN; }
+    public long getISBN() { return ISBN; }
+    public void setISBN(long ISBN) { this.ISBN = ISBN; }
 
-    public Book(int id, String title, Author author, List<String> genres, List<String> subjects, int ISBN, String notes) {
+    public Book(int id, String title, Author author, List<String> genres, List<String> subjects, long ISBN, String notes) {
         this.id = id;
         this.title = title;
         this.author = author;
